lambda_requirements.txt
aws_secretsmanager_caching==1.1.1.5
awswrangler==2.18.0
pandas
pydevd_pycharm
python_dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redshift_connector==2.0.909
s3fs==2022.10.0
setuptools==63.4.1
tomli==2.0.1
